requirements.txt
aiogram==3.11.0
Jinja2==3.1.3
python-dotenv==1.0.1
yookassa==2.4.0
apscheduler==3.10.4
